+++ client/src/components/chat/layout/ChatFooter.tsx
@@ -1,24 +1,20 @@
-import { Suspense, lazy, useState } from "react";
+import { useState } from "react";
 
-import type { EmojiClickData, Theme } from "emoji-picker-react";
-import { Send, Smile } from "lucide-react";
+import type { EmojiClickData } from "emoji-picker-react";
+import { Send } from "lucide-react";
 
+import EmojiPickerButton from "@/components/common/EmojiPickerButton";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
-import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
 import { SheetFooter } from "@/components/ui/sheet";
 
 import { useKeyboardShortcut } from "@/hooks/common/useKeyboardShortcut";
 
 import { useChatStore } from "@/store/useChatStore";
-import { useMediaStore } from "@/store/useMediaStore";
-
-const EmojiPicker = lazy(() => import("emoji-picker-react"));
 
 export default function ChatFooter() {
   const [message, setMessage] = useState<string>("");
   const { sendMessage } = useChatStore();
-  const isMobile = useMediaStore((state) => state.isMobile);
 
   const handleEmojiClick = (emojiData: EmojiClickData) => {
     setMessage((prev) => prev + emojiData.emoji);
@@ -46,30 +42,7 @@ export default function ChatFooter() {
           onChange={(e) => setMessage(e.target.value)}
           className="border-none shadow-none bg-transparent focus-visible:ring-0 focus-visible:ring-offset-0 h-8 px-0 text-sm"
         />
-        {!isMobile && (
-          <Popover>
-            <PopoverTrigger asChild>
-              <Button
-                size="icon"
-                variant="ghost"
-                className="rounded-lg shrink-0 h-8 w-8 text-muted-foreground hover:text-foreground"
-              >
-                <Smile className="!h-6 !w-6" />
-              </Button>
-            </PopoverTrigger>
-            <PopoverContent side="top" align="end" className="w-auto border-none p-0 shadow-none">
-              <Suspense fallback={<div style={{ width: 320, height: 400 }} />}>
-                <EmojiPicker
-                  onEmojiClick={handleEmojiClick}
-                  theme={"auto" as Theme}
-                  lazyLoadEmojis
-                  width={320}
-                  height={400}
-                />
-              </Suspense>
-            </PopoverContent>
-          </Popover>
-        )}
+        <EmojiPickerButton onEmojiClick={handleEmojiClick} />
         <Button
           size="icon"
           className="rounded-lg shrink-0 bg-primary hover:bg-[#2ECC71] text-white h-8 w-8 transition-colors"
